zinovia-backend/pyproject.toml
[project]
name = "zinovia-backend"
version = "0.1.0"
description = "FastAPI backend for Zinovia onboarding chatbot and services API."
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "httpx==0.26.0",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "pydantic[email]==2.5.3",
    "google-auth==2.36.0",
    "sqlalchemy==2.0.25",
    "psycopg2-binary==2.9.9",
    "email-validator==2.1.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "python-dotenv==1.0.0",
    "slowapi==0.1.9",
    "structlog==24.1.0",
    "alembic==1.14.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

